Fix missmatch between label data and output and further minor issues
This commit brings the creation of the model back to the train.ipynb file. It is more convenient for testing purposes to have it at the same place. Also the last layers have been changed so that the output is a one channel image as the label data. The loss function has been changed to BinaryCrossentropy due to the fact that only two classes are considered. Furthermore minor bugs have been fixed.