Add a testing section to the script
This commit adds a testing section to the script which purpose is to show the results of the training. The activation function of the last layer has been switched back to sigmoid and the parameter from_logits in the loss function was set to false. The metrics of the model was set to to binary crossentropy. After this fixes, the training performs as expected.