Organize the code of train.ipynb into corresponding files
This commit provides an early refactoring step organizing the code of train.ipynb into categories in order to be able to reuse the code for means of prediction and transfer learning etc. Also a first attempt to describe the code is made. Also learning parameters are specified in a json file rather than in the code (as far as possible).