Add function which allows to freeze layers by name
This commit adds a function which enables the user to freeze layers by their name rather than just by their index. Differentiating between freezing by index or freezing by name also the function names were set accordingly. The parameter file now also consists of a variable which is suppose to be an array of strings represting the names of the layers.